chore: refactor naming and update styles
