+++ packages/frontend/src/components/LoginForm/SgidLoginSection.tsx
@@ -5,10 +5,12 @@ import {
   Box,
   Divider,
   Flex,
+  Image,
   Link,
   Text,
 } from '@chakra-ui/react'
 import { Button, Infobox } from '@opengovsg/design-system-react'
+import singpassLogo from 'assets/singpass-logo.svg'
 import { SGID_CHECK_ELIGIBILITY_URL } from 'config/urls'
 import { generateSgidAuthUrl } from 'helpers/sgid'
 
@@ -47,7 +49,7 @@ export default function SgidLoginSection(): JSX.Element {
         <Divider />
         <AbsoluteCenter>
           <Box bg="white" p={3}>
-            <Text textStyle="subhead-1">or</Text>
+            <Text textStyle="subhead-1">OR</Text>
           </Box>
         </AbsoluteCenter>
       </Box>
@@ -63,11 +65,12 @@ export default function SgidLoginSection(): JSX.Element {
         <Button
           // isFullWidth a bit ugly
           width="full"
+          variant="outline"
           mb={2}
           onClick={handleSgidLogin}
           isLoading={isRedirectingToSgid}
         >
-          Log in with SingPass
+          Log in with <Image src={singpassLogo} mb={-0.5} h={5} /> app
         </Button>
         <Text>
           Can my agency use this? Check{' '}
